requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-extensions==3.2.3
gunicorn==21.2.0

django-browser-reload
PyJWT==2.8.0
django-cors-headers==4.3.1

# for data and time manipulation
python-dateutil==2.8.2
